test structure bonding
