+++ matminer/featurizers/structure/tests/base.py
@@ -1,0 +1,71 @@
+import copy
+
+from pymatgen.core import Structure, Lattice, Species
+from pymatgen.util.testing import PymatgenTest
+
+
+class StructureFeaturesTest(PymatgenTest):
+    def setUp(self):
+        self.diamond = Structure(
+            Lattice([[2.189, 0, 1.264], [0.73, 2.064, 1.264], [0, 0, 2.528]]),
+            ["C0+", "C0+"],
+            [[2.554, 1.806, 4.423], [0.365, 0.258, 0.632]],
+            validate_proximity=False,
+            to_unit_cell=False,
+            coords_are_cartesian=True,
+            site_properties=None,
+        )
+        self.diamond_no_oxi = Structure(
+            Lattice([[2.189, 0, 1.264], [0.73, 2.064, 1.264], [0, 0, 2.528]]),
+            ["C", "C"],
+            [[2.554, 1.806, 4.423], [0.365, 0.258, 0.632]],
+            validate_proximity=False,
+            to_unit_cell=False,
+            coords_are_cartesian=True,
+            site_properties=None,
+        )
+        self.nacl = Structure(
+            Lattice([[3.485, 0, 2.012], [1.162, 3.286, 2.012], [0, 0, 4.025]]),
+            ["Na1+", "Cl1-"],
+            [[0, 0, 0], [2.324, 1.643, 4.025]],
+            validate_proximity=False,
+            to_unit_cell=False,
+            coords_are_cartesian=True,
+            site_properties=None,
+        )
+        self.cscl = Structure(
+            Lattice([[4.209, 0, 0], [0, 4.209, 0], [0, 0, 4.209]]),
+            ["Cl1-", "Cs1+"],
+            [[2.105, 2.1045, 2.1045], [0, 0, 0]],
+            validate_proximity=False,
+            to_unit_cell=False,
+            coords_are_cartesian=True,
+            site_properties=None,
+        )
+        self.ni3al = Structure(
+            Lattice([[3.52, 0, 0], [0, 3.52, 0], [0, 0, 3.52]]),
+            [
+                "Al",
+            ]
+            + ["Ni"] * 3,
+            [[0, 0, 0], [0.5, 0.5, 0], [0.5, 0, 0.5], [0, 0.5, 0.5]],
+            validate_proximity=False,
+            to_unit_cell=False,
+            coords_are_cartesian=False,
+            site_properties=None,
+        )
+        self.sc = Structure(
+            Lattice([[3.52, 0, 0], [0, 3.52, 0], [0, 0, 3.52]]),
+            ["Al"],
+            [[0, 0, 0]],
+            validate_proximity=False,
+            to_unit_cell=False,
+            coords_are_cartesian=False,
+        )
+        self.bond_angles = range(5, 180, 5)
+
+        diamond_copy = copy.deepcopy(self.diamond)
+        diamond_copy.replace_species(
+            {Species("C", 0.0): {Species("C", 0.0): 0.99, Species("Si", 0.0): 0.01}}
+        )
+        self.disordered_diamond = diamond_copy
